+++ packages/dara-core/dara/core/interactivity/actions.py
@@ -55,6 +55,7 @@ from dara.core.base_definitions import (
 )
 from dara.core.base_definitions import DaraBaseModel as BaseModel
 from dara.core.interactivity.data_variable import DataVariable
+from dara.core.interactivity.state_variable import StateVariable
 from dara.core.internal.download import generate_download_code
 from dara.core.internal.registry_lookup import RegistryLookup
 from dara.core.internal.utils import run_user_handler
@@ -1456,6 +1457,11 @@ class action:
         dynamic_kwargs: Dict[str, AnyVariable] = {}
         static_kwargs: Dict[str, Any] = {}
         for key, kwarg in all_kwargs.items():
+            if isinstance(kwarg, StateVariable):
+                raise ValueError(
+                    'StateVariable cannot be used as input to actions. '
+                    "StateVariables are internal variables for tracking DerivedVariable ephemeral client state shouldn't be used as action payloads."
+                )
             if isinstance(kwarg, AnyVariable):
                 dynamic_kwargs[key] = kwarg
             else: